agregada tercera tienda, Testeo de calculo de distancia agregado

+++ app/src/main/java/com/example/carddotsearcher/model/InventoryItem.kt
@@ -3,6 +3,6 @@ package com.example.carddotsearcher.model
 
 data class InventoryItem(
     val cardName: String,
-    val price: Double, // Usamos Double para representar el precio (ej: 10.99)
+    val price: Int, // Usamos Double para representar el precio (ej: 10.99)
     val stock: Int     // El stock de esta carta específica en la tienda
 )
